Implemented Line intersection.
Plan is that we create lines from bounding boxes. If any of these lines intersect, we have a collision. If at least 2 lines intersect, we detect an object with Radar/Camera. If there are no intersecting points, we have to check whether the object is inside the triangle (camera only, crash detection should automatically disable any movement). This is a TODO for the next commit.